+++ chore/_xtool/llcppsymg/_cmptest/symbol_test/llgo.expect
@@ -1,4 +1,21 @@
 #stdout
+=== Test GenDylibPaths ===
+Test case: Lua library
+Input: -L/opt/homebrew/lib -llua -lm
+Output: [/opt/homebrew/lib/liblua.dylib /opt/homebrew/lib/libm.dylib]
+
+Test case: SQLite library
+Input: -L/opt/homebrew/opt/sqlite/lib -lsqlite3
+Output: [/opt/homebrew/opt/sqlite/lib/libsqlite3.dylib]
+
+Test case: INIReader library
+Input: -L/opt/homebrew/Cellar/inih/58/lib -lINIReader
+Output: [/opt/homebrew/Cellar/inih/58/lib/libINIReader.dylib]
+
+Test case: No valid library
+Input: -L/opt/homebrew/lib
+Error: failed to parse pkg-config output: -L/opt/homebrew/lib
+
 === Test GetCommonSymbols ===
 
 Test Case: Lua symbols
